kernel: timeout: fix "not in list" predication in timeout handling
Issue description by Andy Ross:

The use of dticks == INACTIVE to tell whether or not a timeout was
already in the list was insufficient.  There is a time period between
the moment a timeout is removed from the list and the end of its handler
where it is not in the list, yet its list node pointers still point into
it.  Doing things like aborting a thread while that is true (which can
be asynchronous too!) would corrupt the list even though all the
operations on it were "atomic".

Check for double-remove conditions (which, again, might be perfectly
OK).

Signed-off-by: Peter A. Bigot <pab@pabigot.com>