boards: arm: particle_*: enable PCB antenna on startup
All three boards use a Skyworks SPDT switch to control whether the
antenna is connected to a PCB antenna or an external u.FL connector.
None of them power up in a state that properly enables an antenna.
Add startup code to configure for the PCB antenna.

Upstream-Status: Denied [#15369, use of gpio driver expected]

Signed-off-by: Peter A. Bigot <pab@pabigot.com>