[DNM][WIP] passim: add k_alarm support
k_alarm is analogous to k_timer but with more flexibility and precision.

Alarms are set using an absolute deadline measured in ticks of the
system clock.  The system clock must be maintained as a 64-bit unsigned
counter that will never overflow.  Most operations involve a deadline
specified as a 32-bit unsigned value.  The deadline is compared against
the low 32 bits of a 64-bit system clock.  As long as the system clock
rate does not exceed 64 MHz the 32-bit counter not wrap for at least one
minute; the 64-bit counter permits 10 years at 54 GHz.

There is no alignment of deadlines to Zephyr's notion of a clock tick
(generally 10 ms, minimum 1 ms).  Consequently deadlines are expressed
and triggered with the full accuracy of the system clock, ignoring
delays from higher-priority interrupts.  Further, alarms do not fire
early due to cycles lost as a result of fractional or truncated clock
ticks.

Deadlines that are at or before the current (32-bit) clock value are
considered to be in the past.  For a 32 KiHz system clock this allows
deadlines up to 18:12:15.999969 H:M:S in the future.  (Alarms with
64-bit deadlines could be supported by an additional data structure.)

Unlike k_timer the k_alarm data structure does not support automatic
rescheduling at a fixed interface.  However, the alarm-specific callback
is permitted to reschedule the alarm.  The original deadline is
available in the k_alarm structure for fixed interval reschedule that
ignores delays in callback invocation.  Alternatively, the live clock
can be read and the alarm scheduled at a fixed duration after the time
of last callback.

Alarms that are scheduled with a delay that has already passed become
ready immediately and are processed in order of scheduling within
deadline.  Late alarms scheduled within an alarm callback will be
processed without delay.

Alarms that are ready or scheduled can be cancelled; whether the
cancellation took affect before the callback was invoked is indicated by
the cancellation return value.

WIP gaps to be addressed include:

* Organize declarations so internal and public API is placed appropriate
  in generated documentation;
* Support userspace;
* Determine whether k_alarm_schedule needs an options parameter to
  control things like whether late-to-set alarms are put on the ready
  list or are rejected.

Upstream-Status: Pending [probably not acceptable]

Signed-off-by: Peter A. Bigot <pab@pabigot.com>