soc: nordic_nrf: nrf51: change default SYS_CLOCKS_PER_SEC
The system clock on Nordic is based on 32 KiHz timer.  Scheduling ticks
requires that deadlines be specified with a timer counter that aligns to
a system clock.  With the Zephyr default 100 clocks-per-sec
configuration this results in 100 ticks every 32700 ticks of the cycle
timer.  This reveals two problems:

* The uptime clock misrepresents elapsed time because it runs 0.208%
  (68/32768) faster than the best available clock;

* Calculation of timer counter compare values often requires an integer
  division and multiply operation to produce a value that's a multiple
  of clock-ticks-per-second.

Integer division on the Cortex-M1 nRF51 is done in software with a
(value-dependent) algorithm with a non-constant runtime that can be
significant.  This can produce missed Bluetooth deadlines as discussed
in issue #14577 and others.

By changing the default divisor to one that evenly divides the 2^15
clock rate the time interrupts are disabled to manage timers is
significantly reduced.

Note that the central_hr configuration described in issue #13610 does
not distinguish latency due to timer management from other
irq_block/spinlock regions, and the maximum observed latency will still
exceed the nominal 10 us allowed maximum.  However this does occur
much less frequently than changing the timer deadline which can happen
multiple times per tick.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>