soc: arm: nordic_nrf: unrevert provide custom busy_wait implementations
This reverts commit bd24b31139c1946bd9fe4c336c697b69608fac5e.

The test cause that motivated reverting the fix to produce accurate busy
waits mis-uses the API by trying to busy-wait for 1 s.  The cycle clock
appears to still be stabilizing during the first one or two seconds of
execution, and sometimes runs slow enough that the elapsed time as it
measures is less than the delay produced using CPU cycles.

In any case the gross errors for sub millisecond delays with k_busy_wait()
are far less acceptable than small errors for delays that should use
k_sleep(), so put the accurate implementation back.

Upstream-Status: Submitted [#15707]

Signed-off-by: Peter A. Bigot <pab@pabigot.com>