DNM EVAL samples: bluetooth: central_hr: add 4 KiHz alarm
This modification is intended to verify that the nrf_rtc_timalarm
implementation satisfies the performance requirements of the Bluetooth
subsystem on nRF51, even when there is a parallel 4 KiHz alarm running.

This application ran for over 24 h with no errors, with the following
latency statistics:

[NOTIFICATION] data 0x20003a53 length 2 ctr delta 4096
[24:04:13.987,640] <inf> bt_ctlr_hci: l: 6, 6, 81; t: 45, 32, 96.

Although the 81 is high it is consistent with runs of similar duration
with the unmodified central_hr at Zephyr 1.14.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>