tests/kernel/{fifo,lifo}: Correct tick slop
These tests had what appeared to be a hard-coded one tick slop added
to the timeout for a fifo to handle the addition of a full tick
(vs. proper rounding) in k_sleep() (which is a separate bug).  But the
timeout side has been tightened with the recent change and doesn't
wait as long as an equivalent sleep anymore (which is a feature), so
the sides don't match.

And it's wrong anyway, as there are two sleeps being accounted for (the
fifo timeout and the sleep in the other thread doing the put), both of
which can misalign.  Do this with the proper tick conversion API, and
use two ticks, not one.

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>