[DNM] drivers/display/ssd1306: support selecting inverse display
To set lit text on a black background using the Character Frame Buffer
device, which only supports MONO10 format displays, it's necessary to
invert the display.  Add the hardware configuration necessary for that,
presuming the application correctly selects the appropriate display
format.

DNM: The correct approach is to support MONO01 in CFB.
Signed-off-by: Peter A. Bigot <pab@pabigot.com>