gpio: temporarily allow control of GPIO API deprecation
Some flags and functions will be deprecated as a result of the
addition of new API, but enabling that deprecation produces copious
errors in code that has not yet been converted.  Allow control of the
warnings, with the default at this time being to disable them.

This commit should be reverted when enough in-tree GPIO use has been
converted to avoid gratuitous warnings from builds and sanitycheck.

Signed-off-by: Peter Bigot <peter.bigot@nordicsemi.no>