drivers: serial: fix include order for mcux lpuart
fsl_lpuart.h references an fsl device register LPUART_Type which has a
field named DATA.  If included after <device.h> the macro DATA that is
defined by linker/sections.h causes this to be parsed as data which is
not a field in that structure.

Change the include order to fix the problem.

Signed-off-by: Peter Bigot <peter.bigot@nordicsemi.no>