FIXUP time_units: allow conversion from ticks to us/ns
Existing API produces a 32-bit result for these conversions, which
will often fall into the `mul_ratio` branch of the generic
implementation.  The lack of a generated wrapper requires 64-bit
calculations and truncation in situations where a 32-bit result is
desired and the caller intentionally expects the result to fit.

Provide the additional converters.

Signed-off-by: Peter Bigot <peter.bigot@nordicsemi.no>