FIXUP gpio: esp32: resolve issues with interrupt support
The ESP32 requires specifying a CPU interrupt to be used for GPIO
interrupt signals.  CPU interrupts can be either level or edge (or
special) triggered, but not both.

Update the Kconfig documentation to describe the impact of chosing one
or another line, and in the GPIO driver exclude requests to configure
interrupts for a mode that is not supported.

Also re-order the interrupt acknowledge so the bit is cleared before
the callback is invoked.  I've confirmed that this does work: the
status register bit is cleared.

Disable both-edge detection as this does not seem to function
correctly.

Signed-off-by: Peter Bigot <peter.bigot@nordicsemi.no>