WIP: dts: nordic: add support for QSPI peripheral
Define the basic peripheral information along with devicetree bindings
for the controller and child devices.

The properties for controller and device are mostly based on the
Nordic peripheral register field names, with values specified to be in
natural units, meaning things like frequencies need to be converted to
tick counts in the driver.

An exception is made for SPI mode, for which properties matching Linux
are used.  These properties should also be used in normal SPI nodes,
once Zephyr evolves to using devicetree properly with SPI.

Other properties are shared with other interfaces to the same sort of
device.

Signed-off-by: Peter Bigot <peter.bigot@nordicsemi.no>