DNM: samples: spi_flash: select QSPI to interface to SPI NOR flash.
The configuration file is adjusted to disable the standard driver, and
the bus it's attached to.

An overlay shows how to remove the existing node and replace it with a
new node.

Note that building this will produce the following warning from dtc:

nrf52840_pca10056.dts.pre.tmp:229.22-236.5: Warning (avoid_unnecessary_addr_size): /soc/spi@40023000: unnecessary #address-cells/#size-cells without "ranges" or child "reg" property
  also defined at nrf52840_pca10056.dts.pre.tmp:531.7-552.3
  also defined at nrf52840_pca10056.dts.pre.tmp:590.7-592.3

This will have to be fixed at some point but doesn't block progress.

The application doesn't build because there's no flash driver for this
interface, but it does generate the bindings.

Signed-off-by: Peter Bigot <peter.bigot@nordicsemi.no>