DNM samples: drivers: spi_flash: switch nrf52840_pca10056 to Nordic QSPI
Adjust the configuration file and overlay to disable the SPI driver
and flash node, and replace it with a QSPI driver and flash node.

Note that building this will produce the following warning from dtc:

nrf52840_pca10056.dts.pre.tmp:229.22-236.5: Warning (avoid_unnecessary_addr_size): /soc/spi@40023000: unnecessary #address-cells/#size-cells without "ranges" or child "reg" property
  also defined at nrf52840_pca10056.dts.pre.tmp:531.7-552.3
  also defined at nrf52840_pca10056.dts.pre.tmp:590.7-592.3

The cause of this must be identified and fixed before this can be
merged.

Signed-off-by: Peter Bigot <peter.bigot@nordicsemi.no>