WIP: drivers: clock_control: nrf_power_clock: refactor mutex support
This is work-in-progress while I tried to understand Zephyr 20796.
It's based on an incomplete understanding, and may not be correct.
This commit message describes some of the concerns and the rationale
for changes.  (Other changes are personal style that removed
distractions from my understanding.)

Use a per-device spinlock rather than global irqlock to protect the
clock reference count and list of startup callbacks.  Functionally
this has little/no effect, but it makes the relationship of the
protection more clear.

Rework the logic so that the mutex is held while changes are being
made to either the reference count or the content of the list.  This
ensures that the two do not get out of sync.

Add protection to the get_status implementation, just because it reads
two fields that are normally protected under a mutex.  This API is
probably wrong, as the use of asynchronous start/stop makes it out of
date before the value gets returned to the caller.

Switch to using standard SYS_SLIST API for iterating over the list.

ToDo: document that the request callback may be called from either
interrupt or thread context, and in either case may be invoked before
the request returns.

When the clock start event is processed move all callbacks out of the
device under mutex, then run through them outside mutex.  This avoids
problems with state management that would arise if any of the
callbacks triggered a change in the clock request state.

Concerns remain about robustness in start-stop or stop-start sequences
where the second operation begins before the first has completed; in
the potential for stopping LFCLK which comments in the driver suggest
would not work; and in the effect of invoking the start-calibration
function while the machine is in a non-idle state.