drivers: clock_control: nrf_power_clock: correct ref count on error
clock_control_async_on() is documented to return an error if a
specific client request is repeated before the previous request is
resolved.  The implementation incremented the reference count before
detecting this problem.  Move the check before the increment.  Also
change the returned error code to match the API documentation.

Signed-off-by: Peter Bigot <peter.bigot@nordicsemi.no>