DNM include/posix: incorporate toolchain-provided time.h header
Some of what's supposed to be in <time.h> was lost because this header
attempts to define everything using more primitive include files.
Instead incorporate the contents from the toolchain.

DNM because no attempt has been made to filter out the duplicate
includes and potentially conflicting defines that were added to work
around not including the toolchain version.