WIP samples: drivers: nrf_timer: assess drift with 1pps signal
Track time using the system timer, SysTick, and a TIMER peripheral,
capturing at external 1pps signals and estimating error.

When HFCLK is on the timer and systick errors are less than 1 ppm.  In
that case synth is the highest accuracy LFCLK source, with XTAL then
RC in decreasing accuracy.

When HFCLK is off the timer and systick errors are the same and are
generally between 3000 and 4000 ppm.  Using the synth clock LFCLK
tracks the HFCLK error; RC reduces to between -30 and -60 ppm (30 ppm
resolution); XTAL between 0 and 30 ppm (longer samples show 25 ppm).

Signed-off-by: Peter Bigot <peter.bigot@nordicsemi.no>