drivers: ethernet: gecko: device config_info content mutated
The device config_info structure is intended to be read-only, and so
the generic pointer stored in struct device is a pointer-to-const.
The gecko driver removes this qualifier when casting the pointer to
the device-specific type.

When the qualifier is preserved the driver does not build, because it
modifies the object content at runtime.

This patch provides the changes necessary to preserve const
correctness.  When applied it produces the following diagnostic.  The
driver must be updated to resolve this.

drivers/i2c/i2c_gecko.c:95:23: error: assignment of member 'freq' in read-only object
   95 |  config->i2cInit.freq = baudrate;

Signed-off-by: Peter Bigot <peter.bigot@nordicsemi.no>