drivers: audio: tlv320dac310x: device config_info content mutated
The device config_info structure is intended to be read-only, and so
the generic pointer stored in struct device is a pointer-to-const.
The tlv320dac310x driver removes this qualifier when casting the pointer to
the device-specific type.  This removal is a violation of MISRA 11.8
as it evokes undefined behavior.

When the qualifier is preserved the driver does not build, because it
modifies the object content at runtime.

https://github.com/pabigot/zephyr/commits/nordic/issue/24745b provides
the changes necessary to preserve const correctness.  When applied it
produces the following diagnostic, among others:

```
/home/buildslave/src/github.com/zephyrproject-rtos/zephyr/drivers/audio/tlv320dac310x.c: In function 'codec_initialize':
/home/buildslave/src/github.com/zephyrproject-rtos/zephyr/drivers/audio/tlv320dac310x.c:81:22: error: assignment of member 'i2c_device' in read-only object
   81 |  dev_cfg->i2c_device = device_get_binding(dev_cfg->i2c_dev_name);
      |                      ^
/home/buildslave/src/github.com/zephyrproject-rtos/zephyr/drivers/audio/tlv320dac310x.c:89:23: error: assignment of member 'gpio_device' in read-only object
   89 |  dev_cfg->gpio_device = device_get_binding(dev_cfg->gpio_dev_name);
```

The driver must be updated to resolve this.

Signed-off-by: Peter Bigot <peter.bigot@nordicsemi.no>