drivers: audio: dma_nios2_msgdma: device config_info content mutated
The device config_info structure is intended to be read-only, and so
the generic pointer stored in struct device is a pointer-to-const.
The dma_nios2_msgdma driver removes this qualifier when casting the pointer to
the device-specific type.  This removal is a violation of MISRA 11.8
as it evokes undefined behavior.

When the qualifier is preserved the driver does not build, because it
modifies the object content at runtime.  Consequently the fix for this
driver has been removed from #25248.

https://github.com/pabigot/zephyr/commits/nordic/issue/24745c provides
the changes necessary to preserve const correctness.  When applied it
produces the following diagnostic, among others:

```
/home/buildslave/src/github.com/zephyrproject-rtos/zephyr/drivers/dma/dma_nios2_msgdma.c: In function 'nios2_msgdma_config':
/home/buildslave/src/github.com/zephyrproject-rtos/zephyr/drivers/dma/dma_nios2_msgdma.c:105:13: error: passing argument 1 of 'k_sem_take' discards 'const' qualifier from pointer target type [-Werror=discarded-qualifiers]
  105 |  k_sem_take(&dev_cfg->sem_lock, K_FOREVER);
      |             ^~~~~~~~~~~~~~~~~~
In file included from ../../../../../../../include/kernel.h:5421,
                 from ../../../../../../../include/init.h:11,
                 from ../../../../../../../include/device.h:22,
                 from /home/buildslave/src/github.com/zephyrproject-rtos/zephyr/drivers/dma/dma_nios2_msgdma.c:7:
zephyr/include/generated/syscalls/kernel.h:746:45: note: expected 'struct k_sem *' but argument is of type 'const struct k_sem *'
  746 | static inline int k_sem_take(struct k_sem * sem, k_timeout_t timeout)
      |                              ~~~~~~~~~~~~~~~^~~
/home/buildslave/src/github.com/zephyrproject-rtos/zephyr/drivers/dma/dma_nios2_msgdma.c:106:24: error: assignment of member 'dma_callback' in read-only object
  106 |  dev_cfg->dma_callback = cfg->dma_callback;
      |                        ^
```

The driver must be updated to resolve this and incorporate the change
from the referenced branch.

Signed-off-by: Peter Bigot <peter.bigot@nordicsemi.no>