bluetooth: mesh: access: avoid C99 features unavailable in C++
C99 designated array initializers are not part of the C++ language,
and are not supported in GNU G++.  Their use in this file does not
appear to be necessary, as a standard initialization has the same
behavior.

Remove the unsupported language feature, and fix the initialization
order to match declaration order (which C++ also requires).

Signed-off-by: Peter Bigot <peter.bigot@nordicsemi.no>