drivers: i2c: Add a 'transfer' method to the slave callback
The current I2C slave-callback functions are fairly low level, tied in
with the way the STM32 I2C driver works. This is natural, since it is
the only driver that implements the API.

However it is a bit painful to work with. It is simpler to pass the
full I2C message info across the interface. This results in a lot less
code and complexity in the implementation.

Add a new slave_transfer() method to the API which does this.

Signed-off-by: Simon Glass <sjg@chromium.org>