tests: i2c: Don't rely on ordering in i2c_slave_api test
At present this test seems to rely on the numbering or ordering of
devices. This doesn't seem to work on posix. Instead of this, read the
data and check whether it matches what is expected. Then set the
expected data accordingly.

This should work regardless of the devices and ordering.

Signed-off-by: Simon Glass <sjg@chromium.org>