FIXUP: use node label to disambiguate devices
There should be no order dependency in this test.  The only potential
issue is that instance numbers may not be consistent between builds.
That is averted by using node labels rather than hard-coded instances
to identify the distinct devices.

In practice this shouldn't matter because the test should not care
which is which.

Signed-off-by: Peter Bigot <peter.bigot@nordicsemi.no>