drivers: emul: Third attempt to create an I2C emulator
An I2C emulator pretends to be an I2C device. A special I2C controller
driver handles finding the emulator for a particular I2C bus/address
and sending traffic to the emulator for processing. No changes are
needed to the driver.

This includes:

- drivers/emul       - new directory for emulators
- i2c_emul_ctlr.c    - I2C controller which uses emulators instead of
                       real hardware
- eeprom_at24_emu.c  - Emulator for Atmel AT24 I2C ERPROM

The test to run is in tests/drivers/i2c/i2c_emul/src/main.c

This is only for discussion as it does not go far enough yet.

Signed-off-by: Simon Glass <sjg@chromium.org>