hackery on I2C emulator
A grotesque butchery intended to validate my concept for how to
associate emulators with devices and use them to run the standard
driver tests for emulated I2C devices.

tests/drivers/eeprom adds an overlay for native_posix that discards
the existing eeprom node (which uses a simulated EEPROM) and defines
an AT24 eeprom on the emulated I2C bus.  The Kconfig settings are
updated to select the AT2X driver instead of the SIMULATED driver.

(I propose to use "emul_<compatible>" as the canonical name for an
emulator supporting a driver for a particular compatible.  New files
and API in this commit may follow that standard.)

A generic emulator registration structure is defined as an analog to
struct init_entry for devices.  It provides the function to be used to
initialize the emulator, a handle that allows emulating controllers to
identify the emulators they need, and a pointer to emulator-specific
configuration data.  These structures are aggregated into an array in
ROM, analogous to the array of device structures.

When the emulating controller (i2c_emul device) is initialized it
walks the list of devices attached to it, matches each device against
its emulator registration record in the above array, and invokes the
corresponding emulator init function providing the parent device.

The emulator (at24) instance initializes itself, then knowing the
parent is an i2c_emul controller calls back through that API to
register itself as a device on the bus.

The pointer to the i2c emulator registration is used by the emulating
controller to pass state to the emulator, as the structure is embedded
within the mutable data of the emulator instance.  This improves type
safety over the void pointer solution proposed initially.

The test to run is tests/drivers/eeprom.

Other changes that seemed appropriate or simplified the above changes:

For the device emulator (at24) the data structures have been
refactored to put more material into the const (ROM) config structure,
with only the material that needs to change into the data structure.

Also the i2c_emul controller has been extended with driver-specific
API to allow emulators that need to know the controller configuration
to fetch it, rather than having every emulator cache a copy whenever
it changes.

tests/drivers/i2c/i2c_emul has been ignored as it is replaced by
tests/drivers/eeprom.  It could be replaced by something that tests
the emulating controller in isolation if that's useful.

Signed-off-by: Peter Bigot <peter.bigot@nordicsemi.no>