hello_world: reimplement as timer test
This repurposes the hello_world sample as a test of various policies
for handling periodic operations.  Multiple timers are created with
the same control flow: on timeout schedule a work item.  Each timer is
rescheduled following a different policy, but all repeat at a nominal
interval of 1 s.

This application will fail usually within the first couple minutes.
When the failure occurs the first timeout in the group has completed,
but the kernel timeout_list is pointing to a timeout which is
registered with dticks=1.  This timeout is not recognized or processed
until the 24-bit 32 KiHz timer laps, stalling the application.

The problem is exacerbated by the timer interrupt handler including a
50 us busywait to simulate doing some non-trivial other operations.
In the unreduced code this was actual work.

I've reproduced this on nrf52840dk_nrf52840 and nrf52dk_nrf52832.  The
failure has not reproduced on nrf51dk_nrf51422.

Signed-off-by: Peter Bigot <peter.bigot@nordicsemi.no>