fix sign issue at user level
The calculation of the next timeout was done in unsigned arithmetic,
which when the latency exceeds the interval wraps making the next
schedule point appear to be far in the future.

This is demonstrated by the "rel sch" periodic instance not appearing
again after the first failure.

Calculate the delay before the next invocation using signed arithmetic.

It needs to be verified that the underlying infrastructure doesn't
still cause this to be transformed into a large unsigned value.

Signed-off-by: Peter Bigot <peter.bigot@nordicsemi.no>