net: tcp2: work around limitations of legacy API
The tcp2 infrastructure is using the legacy delayed work API, and
relies heavily on the transient state indicated by an estimate of
delayed time remaining to determine whether a delayed work item is
still active.  While the wrappers for this work in most cases, one use
is unsanctioned: directly submitting a delayed item to the work queue
to process immediately, using the non-public fields of the delayed
work structure.

The solution for this specific need in the new API is to use a
schedule (rather than reschedule) operation, which leaves any previous
timer unchanged but allows immediate submission if the work is idle.

Put in a hack that also uses the non-public fields of the delayed work
structure to implement the required behavior.  The complete fix
requires replacing all use of k_delayed_work in this module with
k_work_delayable, leveraging the new functionality of the API to avoid
having to guess about the true state of a work item based on its
transient timer or flag states.

Signed-off-by: Peter Bigot <peter.bigot@nordicsemi.no>