lib: cbprintf: refactor to expose some internals
There are use cases where it's necessary to inspect the conversion
specifier and the argument value, e.g. to make copies of arguments
when the formatting must be done in a context other than the call
site.  Expose the structure that encodes the conversion specification,
and the structure that captures argument values, along with supporting
constants necessary to interpret their contents.

Signed-off-by: Peter Bigot <peter.bigot@nordicsemi.no>