WIP: lib: cbprintf: prototype API to generalize format extraction
This was intended to implement functionality that walks the format and
arguments, allowing argument values to be mutated by a callback, and
then stored in a location where they could be processed later.

It fails because of limitations of C and the processing of va_list
objects.  Use of va_get() mutates the va_list object which is
nominally passed by value, so all extractions from the object need to
be done in the same function body.

Naively one would expect that extractions could be done with a
separate function by passing a pointer to the va_list object, so
mutations are visible in both the called and calling function.  But
this doesn't work on x86_64 because a `va_list` is not a pointer, or a
structure; it's an array.

See details at https://stackoverflow.com/a/8048892.  Some potential
workarounds are discussed there, but the overall conclusion is that
pulling arguments off a va_list from different functions is not
portable.

Signed-off-by: Peter Bigot <peter.bigot@nordicsemi.no>