PROTOTYPE: use C11 _Generic to support polymorphic calls
Multiple work queue operations need to be supported on both standard
work items and delayable work items.  A move to requiring the last
component of the function name to be a verb introduces ambiguity: while

  k_work_init(&w)

is clearly initializing a `k_work` the other variant:

  k_work_delayable_init(&w)

could be initializing a `k_work_delayable` (as intended) or doing a
delayed initialization of a `k_work`.

Similar issues arise for `k_work_cancel`, `k_work_busy`,
`k_work_pending`, `k_work_flush`, ...

This commit prototypes use of C11 Generics to allow a common visible
API `k_work_init()` that is redirected to functions specific to the
type of work item involved.

Signed-off-by: Peter Bigot <peter.bigot@nordicsemi.no>